refactorizando el código
