--- client/room.html
@@ -1,147 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>ChatSocket</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
-  <style>
-    #messages {
-      height: 60vh;
-      overflow-y: scroll;
-    }
-
-  </style>
-</head>
-
-<body>
-  <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-    <div class="container">
-      <a class="navbar-brand" href="#">ChatSocket</a>
-      <div class="navbar-nav">
-        <a class="nav-link" href="/chat">Chat</a>
-        <a class="nav-link" href="logout.html">Logout</a>
-      </div>
-    </div>
-  </nav>
-
-  <div class="container mt-5">
-    <h1 id="room-name" class="text-center mb-5">Room Name</h1>
-
-    <div class="row justify-content-center">
-      <div class="col-md-8">
-        <div id="messages">
-          <!-- Messages will be displayed here -->
-        </div>
-
-        <form id="message-form" class="mt-3">
-          <div class="row">
-            <div class="col-md-9">
-              <input type="text" class="form-control" id="message-input" placeholder="Type your message...">
-            </div>
-
-            <div class="col-md-3">
-              <button type="submit" class="btn btn-primary w-100">Send</button>
-            </div>
-          </div>
-        </form>
-
-        <script src="http://localhost:4000/socket.io/socket.io.js"></script>
-        <script>
-          const token = localStorage.getItem( 'token' );
-          const socket = io( 'http://localhost:4000', {
-            auth: {
-              token: token
-            }
-          } );
-
-          const messagesContainer = document.getElementById( "messages" );
-          const messageForm = document.getElementById( "message-form" );
-          const messageInput = document.getElementById( "message-input" );
-          const roomName = localStorage.getItem( "roomName" );
-
-
-
-          // Join the room and get the room information from the server
-          socket.emit( "join-room", roomName, ( room ) => {
-            console.log( 'room info', room );
-            // Update the room name in the HTML
-            const roomNameElement = document.getElementById( "room-name" );
-            roomNameElement.textContent = room.name;
-
-            // Display the messages in the room
-            room.messages.forEach( ( message ) => {
-              addMessage( message );
-            } );
-          } );
-
-          messageForm.addEventListener( "submit", ( event ) => {
-            event.preventDefault();
-            const message = messageInput.value;
-            console.log( 'sending message', message );
-            data = {
-              room: roomName,
-              message: message,
-              username: localStorage.getItem( "username" )
-            };
-            socket.emit( "send-message", data );
-            console.log( 'message sent', data );
-
-            addMessage( data );
-            messageInput.value = "";
-          } );
-
-          socket.on( "send-message", ( data ) => {
-            console.log( 'response from server', data );
-            addMessage( data );
-          } );
-
-          function addMessage ( data ) {
-            const messageClass = data.isCurrentUser ? "background-color: #007bff; color: #fff;" : "background-color: #c1e1c1; color: #000;";
-            const userClass = "color: #000;";
-            const token = data.isCurrentUser ? localStorage.getItem( "token" ) : data.token;
-
-            const messageHTML = `
-    <div class="my message p-2 rounded mb-2" style="${ messageClass }">
-      <p class="mb-0">${ data.message }</p>
-      <div class="d-flex justify-content-end">
-        <small style="${ userClass }">${ data.user.username }</small>
-      </div>
-    </div>
-  `;
-            messagesContainer.insertAdjacentHTML( "beforeend", messageHTML );
-            messagesContainer.scrollTop = messagesContainer.scrollHeight;
-          }
-          // Debugging instructions
-          socket.on( 'connect', () => {
-            console.log( 'Connected to server' );
-          } );
-
-          socket.on( 'disconnect', () => {
-            console.log( 'Disconnected from server' );
-          } );
-
-          socket.on( 'error', ( error ) => {
-            console.error( 'Socket error', error );
-          } );
-
-
-        </script>
-        <script>
-          function updateRoomTitle () {
-            const roomNameElement = document.getElementById( "room-name" );
-            const roomName = localStorage.getItem( "roomName" );
-            roomNameElement.textContent = roomName;
-          }
-
-          // Llamamos a la función para actualizar el título de la sala
-          updateRoomTitle();
-        </script>
-      </div>
-    </div>
-  </div>
-</body>
-
-</html>
